requirements.txt
python-telegram-bot
xlrd
pandas
plotly.express
kaleido
python-decouple
numpy
datetime
#gzip
#shutil